+++ examples/basic_image2.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 require 'glimmer-dsl-libui'
 
 include Glimmer
@@ -7,7 +5,7 @@ include Glimmer
 window('Basic Image', 96, 96) {
   area {
     on_draw do |area_draw_params|
-      image(File.expand_path('../icons/glimmer.png', __dir__), 96, 96)
+      image(File.expand_path('../icons/glimmer.png', __dir__), height: 96)
     end
   }
 }.show