+++ examples/basic_image4.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 require 'glimmer-dsl-libui'
 
 include Glimmer
@@ -9,7 +7,6 @@ window('Basic Image', 96, 96) {
     on_draw do |area_draw_params|
       image {
         file File.expand_path('../icons/glimmer.png', __dir__)
-        width 96
         height 96
       }
     end